Added  `WaveFile::addTrack()`
This takes a 2D array of PianoNote. The intention is to construct the WaveFile with one track (ex: the notes in the treble clef) and call addTrack (ex: for bass clef notes).

addTrack REQUIRES the newTrack to have the same duration as that which was used to construct the WaveFile, but this is fairly reasonable, I think.